AG-10044 - Refactor grid scss modules names
