+++ packages/ag-charts-website/src/components/site-header/HeaderNav.tsx
@@ -1,7 +1,7 @@
 import type { MenuItem } from '@ag-grid-types';
 import { SITE_BASE_URL } from '@constants';
 import styles from '@design-system/modules/HeaderNav.module.scss';
-import gridStyles from '@design-system/modules/gridSiteHeader.module.scss';
+import gridStyles from '@design-system/modules/SiteHeader.module.scss';
 import MenuIcon from '@images/inline-svgs/menu-icon.svg?react';
 import { pathJoin } from '@utils/pathJoin';
 import classnames from 'classnames';
