+++ packages/ag-charts-website/src/components/site-header/SiteHeader.astro
@@ -4,7 +4,7 @@ import { Icon } from '@components/icon/Icon';
 import { HeaderNav } from './HeaderNav';
 import { getEntry } from 'astro:content';
 import { FRAMEWORKS } from '@constants';
-import gridStyles from '@design-system/modules/gridSiteHeader.module.scss';
+import gridStyles from '@design-system/modules/SiteHeader.module.scss';
 import { SiteLogo } from './SiteLogo';
 import Search from '../search/Search';
 import { getFrameworkFromPath } from '../../features/docs/utils/urlPaths';
