+++ packages/ag-charts-website/src/design-system/modules/Search.module.scss
