+++ packages/ag-charts-website/src/design-system/modules/SearchBox.module.scss
