+++ packages/ag-charts-website/src/design-system/modules/SearchResult.module.scss
